requirements.txt
# Google ADK and AI dependencies
google-adk[a2a]>=1.22.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
uvicorn>=0.27.0

# Agent orchestration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
asyncio>=3.4.3
aiohttp>=3.9.0

# Infrastructure generation
jinja2>=3.1.2
pyyaml>=6.0.1

# Security scanning (for Security Agent)
safety>=2.3.0

# Observability
prometheus-client>=0.19.0

# CLI utilities
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.2

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
